GET request to the "/Identities" route of Admin API returns http status 500 (#144)
* fix: cannot resolve UserManager in the Identities repository

* Revert "fix: cannot resolve UserManager in the Identities repository"

This reverts commit c45304bc1136437e71aa8f08c302712b5914a195.

* Revert "Revert "fix: cannot resolve UserManager in the Identities repository""

This reverts commit 3af498c089a2b2e59f32fc01f84e5d0fd7e4bb1a.

* fix: consolidate AddCustomIdentity() into BuildingBlocks.API
