+++ Admin.API/Extensions/IServiceCollectionExtensions.cs
@@ -122,6 +122,7 @@ public static class IServiceCollectionExtensions
 
         return services;
     }
+
     private static object GetPropertyValue(object source, string propertyPath)
     {
         foreach (var property in propertyPath.Split('.').Select(s => source.GetType().GetProperty(s)))
