+++ BuildingBlocks/src/BuildingBlocks.API/BuildingBlocks.API.csproj
@@ -24,6 +24,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\..\..\Modules\Devices\src\Devices.Domain\Devices.Domain.csproj" />
+		<ProjectReference Include="..\..\..\Modules\Devices\src\Devices.Infrastructure\Devices.Infrastructure.csproj" />
 		<ProjectReference Include="..\BuildingBlocks.Application.Abstractions\BuildingBlocks.Application.Abstractions.csproj" />
 		<ProjectReference Include="..\BuildingBlocks.Application\BuildingBlocks.Application.csproj" />
 		<ProjectReference Include="..\Tooling\Tooling.csproj" />
